data/requirements.txt
gamsapi[transfer] == 50.1.0
pandas